requirements.txt
requests>=2.31.0
graphviz>=0.20.1
pyyaml>=6.0.1
python-dotenv>=1.0.0
